image-rs: add feature gate for verity
