+++ image-rs/Cargo.toml
@@ -15,7 +15,7 @@ async-trait.workspace = true
 attestation_agent = { path = "../attestation-agent/lib", optional = true }
 base64.workspace = true
 cfg-if = { workspace = true, optional = true }
-devicemapper = "0.33.5"
+devicemapper = { version =  "0.33.5", optional = true }
 dircpy = { version = "0.3.12", optional = true }
 flate2 = "1.0"
 fs_extra = { version = "1.2.0", optional = true }
@@ -119,3 +119,5 @@ snapshot-unionfs = ["nix", "dircpy", "fs_extra"]
 getresource = [ "lazy_static", "cfg-if" ]
 
 nydus = ["lazy_static", "nydus-api", "nydus-service"]
+
+verity = ["devicemapper"]
